backend/requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
Authlib==1.3.2
bcrypt==4.1.2
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pillow==11.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.10.0
pydantic-settings==2.6.0
pydantic_core==2.27.0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.12
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
typing_extensions==4.15.0
uvicorn==0.32.0
watchfiles==1.1.1
websockets==15.0.1
